検索結果作成

+++ entity/src/main/java/com/example/model/CacheBookInfo.kt
@@ -14,7 +14,7 @@ data class CacheBookInfo(
     @ColumnInfo val description: String,
     @ColumnInfo val pageCount: Int,
     @ColumnInfo val categories: String,     // カンマ区切り
-    @ColumnInfo val averageRating: Double?,
+    @ColumnInfo val averageRating: Int?,
     @ColumnInfo val ratingCount: Int,
     @ColumnInfo val image: String,
     @ColumnInfo val language: String,
